requirements.txt
python-dotenv
Flask==2.0.2
openai
pylint
pytest